requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
asn1crypto==1.5.1
    # via scramp
certifi==2025.10.5
    # via requests
chardet==4.0.0
    # via requests
exceptiongroup==1.3.0
    # via pytest
greenlet==3.2.4
    # via sqlalchemy
idna==2.10
    # via requests
iniconfig==2.3.0
    # via pytest
numpy==1.23.4
    # via
    #   -r requirements.in
    #   pandas
packaging==25.0
    # via pytest
pandas==2.3.3
    # via -r requirements.in
pg8000==1.31.5
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.3
    # via pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-mock
pytest-mock==3.15.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   pg8000
python-dotenv==1.1.1
    # via pydantic-settings
pytz==2025.2
    # via pandas
requests==2.25.1
    # via -r requirements.in
scramp==1.4.6
    # via pg8000
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.5.post1
    # via -r requirements.in
tomli==2.3.0
    # via pytest
typing-extensions==4.15.0
    # via
    #   exceptiongroup
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==1.26.20
    # via requests
